chore: rename data-source to snake case
