+++ interapp-backend/db/init.ts
@@ -1,4 +1,4 @@
-import AppDataSource from '@db/data-source';
+import AppDataSource from '@db/data_source';
 
 AppDataSource.initialize()
   .then(() => {
